normed, fixed texture leaks, few last fixes to do then rdy
