+++ inc/ansi_hex_codes.h
@@ -1,0 +1,62 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ansi_hex_codes.h                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/16 12:02:55 by abied-ch          #+#    #+#             */
+/*   Updated: 2024/01/16 12:03:04 by abied-ch         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef ANSI_HEX_CODES_H
+# define ANSI_HEX_CODES_H
+
+# define GREEN		"\e[92;5;118m"
+# define RED 		"\e[91m"
+# define YELLOW 	"\e[93;5;226m"
+# define GRAY		"\e[33;2;37m"
+# define RESET 		"\e[0m"
+# define CURSIVE 	"\e[33;3m"
+# define BOLD 		"\e[1m"
+
+# define HEXA_RED	0x00FF0000
+# define HEXA_GREEN	0x0000FF00
+# define HEXA_BLUE	0x000000FF
+# define HEXA_WHITE	0x00FFFFFF
+# define HEXA_BLACK	0x00000000
+# define HEXA_YELLOW 0x00FFFF00
+# define HEXA_GRAY	0x00A9A9A9
+# define HEXA_PURPLE 0x00FF00FF
+# define HEXA_CYAN	0x0000FFFF
+# define HEXA_ORANGE 0x00FFA500
+# define HEXA_PINK	0x00FFC0CB
+# define HEXA_BROWN	0x00A52A2A
+# define HEXA_GOLD	0x00FFD700
+# define HEXA_SILVER 0x00C0C0C0
+# define HEXA_BRONZE 0x00CD7F32
+# define HEXA_LIME	0x00BFFF00
+# define HEXA_OLIVE	0x00808000
+# define HEXA_NAVY	0x00008080
+# define HEXA_TEAL	0x00008080
+# define HEXA_MAROON 0x00800000
+# define HEXA_INDIGO 0x004B0082
+# define HEXA_TURQUOISE 0x0040E0D0
+# define HEXA_VIOLET 0x00EE82EE
+# define HEXA_BEIGE 0x00F5F5DC
+# define HEXA_AQUA 0x0000FFFF
+# define HEXA_CORAL 0x00FF7F50
+# define HEXA_SKY_BLUE 0x0087CEEB
+# define HEXA_MINT 0x00F5FFFA
+# define HEXA_LAVENDER 0x00E6E6FA
+# define HEXA_TAN 0x00D2B48C
+# define HEXA_SALMON 0x00FA8072
+# define HEXA_KHAKI 0x00F0E68C
+# define HEXA_CRIMSON 0x00DC143C
+# define HEXA_ORCHID 0x00DA70D6
+# define HEXA_SPACE_BLUE 0x002147
+# define HEXA_MIDNIGHT_BLUE 0x191970
+# define HEXA_BORDEAUX 0x800020
+
+#endif
