--- minilibx-mac/mlx_png.h
@@ -1,4 +1,0 @@
-
-
-
-void    *mlx_png_file_to_image(void *xvar, char *file, int *width, int *height);
