+++ src/raycaster/initialize.c
@@ -6,50 +6,55 @@
 /*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/17 12:33:56 by abied-ch          #+#    #+#             */
-/*   Updated: 2023/12/21 14:44:19 by abied-ch         ###   ########.fr       */
+/*   Updated: 2024/01/16 12:18:09 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/raycast.h"
 
-void	set_data_view(t_data *data)
+void	get_texture_data(t_txtr *texture, t_data *data, char *texture_path)
 {
-	int		x;
-	float	x_scale;
-	float	y_scale;
-	int		map_x;
-	int		map_y;
-
-	x_scale = (float)SCREEN_WIDTH / data->map_width;
-	y_scale = (float)SCREEN_HEIGHT / data->map_height;
-	x = -1;
-	while (++x < data->map_height * data->map_width)
-	{
-		data->view_dir = isdirection(data->map[x]);
-		if (data->view_dir)
-		{
-			map_x = (x % data->map_width);
-			map_y = (x / data->map_height);
-			data->player.x_pos = map_x;
-			data->player.y_pos = map_y;
-			data->player.x_screen_pos = map_x * x_scale + x_scale / 2;
-			data->player.y_screen_pos = map_y * y_scale + y_scale / 2;
-			return (set_directions(data));
-		}
-	}
+	texture->img = mlx_xpm_file_to_image(data->mlx.mlx, texture_path,
+			&texture->width, &texture->height);
+	if (texture->img == NULL)
+		exit_failure(data, "Error\nimage initialization failed");
+	texture->addr = mlx_get_data_addr(texture->img, &texture->bpp,
+			&texture->l_l, &texture->endian);
 }
 
-float	min(float a, float b)
+void	initialize_textures(t_data *data)
 {
-	if (a < b)
-		return (a);
-	return (b);
+	get_texture_data(&data->brick, data, "src/textures/brick.xpm");
+	get_texture_data(&data->stone, data, "src/textures/stone.xpm");
+	get_texture_data(&data->pepe, data, "src/textures/pepe.xpm");
+	get_texture_data(&data->wood, data, "src/textures/wood.xpm");
+	get_texture_data(&data->grass, data, "src/textures/grass.xpm");
+	get_texture_data(&data->metal, data, "src/textures/metal.xpm");
+	get_texture_data(&data->walltest, data, "src/textures/walltest.xpm");
+	get_texture_data(&data->wall1, data, "src/textures/Wall1.xpm");
+	get_texture_data(&data->wall2, data, "src/textures/Wall2.xpm");
+	get_texture_data(&data->wall3, data, "src/textures/Wall3.xpm");
+	get_texture_data(&data->wall4, data, "src/textures/Wall4.xpm");
+	get_texture_data(&data->wall5, data, "src/textures/Wall5.xpm");
+	get_texture_data(&data->wall6, data, "src/textures/Wall6.xpm");
+	get_texture_data(&data->wall7, data, "src/textures/Wall7.xpm");
+	get_texture_data(&data->wall8, data, "src/textures/Wall8.xpm");
+}
+
+void	new_image(t_data *data)
+{
+	if (data->mlx.img)
+		mlx_destroy_image(data->mlx.mlx, data->mlx.img);
+	data->mlx.img = mlx_new_image(data->mlx.mlx,
+			data->win_width, data->win_height);
+	data->img.addr = mlx_get_data_addr(data->mlx.img, &data->img.bpp,
+			&data->img.l_l, &data->img.endian);
+	raycast(data);
+	mlx_put_image_to_window(data->mlx.mlx, data->mlx.win, data->mlx.img, 0, 0);
 }
 
 void	initialize_data(t_data *data)
 {
-	int	player_pos;
-
 	data->mlx.mlx = NULL;
 	data->mlx.win = NULL;
 	data->mlx.img = NULL;
@@ -69,7 +74,6 @@ void	initialize_data(t_data *data)
 	data->player.y_dir = sin(data->player.angle) * 5;
 	data->map_size = data->map_height * data->map_width;
 	data->min_distance = 0;
-	player_pos = map_get_player_pos(data->map);
-	data->player.x_pos = player_pos % data->map_width;
-	data->player.y_pos = (float)player_pos / data->map_width;
+	data->player.x_pos = map_get_player_pos(data->map) % data->map_width;
+	data->player.y_pos = (float)map_get_player_pos(data->map) / data->map_width;
 }
