+++ src/raycaster/math_utils.c
@@ -6,20 +6,31 @@
 /*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/21 10:29:33 by abied-ch          #+#    #+#             */
-/*   Updated: 2023/12/21 10:43:14 by abied-ch         ###   ########.fr       */
+/*   Updated: 2024/01/16 12:45:34 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/raycast.h"
 
-float	normalize_angle(float angle, float increment)
+bool	collision(t_data *data, float new_x, float new_y)
 {
-	angle += increment;
-	if (angle < 0)
-		angle += 2 * PI;
-	else if (angle > 2 * PI)
-		angle -= 2 * PI;
-	return (angle);
+	int	map_pos;
+
+	map_pos = (int)new_y * data->map_width + (int)new_x;
+	if (map_pos < 0 || map_pos > data->map_width * data->map_height)
+		return (true);
+	if (data->map[map_pos] != 1)
+		return (false);
+	return (true);
+}
+
+void	normalize_angle(float *angle, float increment)
+{
+	*angle += increment;
+	if (*angle < 0)
+		*angle += 2 * PI;
+	else if (*angle > 2 * PI)
+		*angle -= 2 * PI;
 }
 
 float	dist(t_line line)
