+++ src/raycaster/minimap.c
@@ -6,7 +6,7 @@
 /*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/17 12:34:11 by abied-ch          #+#    #+#             */
-/*   Updated: 2023/12/21 12:56:19 by abied-ch         ###   ########.fr       */
+/*   Updated: 2024/01/16 12:25:40 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,16 +47,6 @@ void	draw_map(t_data *data)
 				continue ;
 		}
 	}
-}
-
-void	init_line(t_line *line, t_data *data, int length)
-{
-	if (!length)
-		length = 1;
-	line->x0 = data->player.x_pos;
-	line->y0 = data->player.y_pos;
-	line->x1 = (data->player.x_pos + data->player.x_dir);
-	line->y1 = (data->player.y_pos + data->player.y_dir);
 }
 
 void	draw_player(t_data *data)
