+++ src/raycaster/texture_utils.c
@@ -6,7 +6,7 @@
 /*   By: abied-ch <abied-ch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/21 10:31:37 by abied-ch          #+#    #+#             */
-/*   Updated: 2024/01/15 21:49:21 by abied-ch         ###   ########.fr       */
+/*   Updated: 2024/01/16 12:43:20 by abied-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,11 +24,25 @@ void	draw_background(t_data *data)
 	{
 		y = -1;
 		while (++y < SCREEN_HEIGHT)
-		if (y < SCREEN_HEIGHT / 2)
-			put_pixel(data, x, y, data->ceiling_color);
-		else
-			put_pixel(data, x, y, data->floor_color);
+		{
+			if (y < SCREEN_HEIGHT / 2)
+				put_pixel(data, x, y, data->ceiling_color);
+			else
+				put_pixel(data, x, y, data->floor_color);
+		}
 	}
+}
+
+void	set_texture(t_data *data, t_txtr *texture)
+{
+	if (data->hit == NORTH)
+		*texture = data->wall1;
+	else if (data->hit == EAST)
+		*texture = data->wall2;
+	else if (data->hit == SOUTH)
+		*texture = data->wall3;
+	else if (data->hit == WEST)
+		*texture = data->wall4;
 }
 
 int	get_pixel(t_txtr *texture, int x, int y)
@@ -43,6 +57,12 @@ int	get_pixel(t_txtr *texture, int x, int y)
 	return (color);
 }
 
+void	set_hit_position(t_data *data)
+{
+	data->hit_pos = fmod(data->hit_pos / 3.79, data->map_width);
+	data->hit_pos /= data->map_width;
+}
+
 void	draw_texture(t_data *data, int x, t_line line, t_txtr *texture)
 {
 	int		color;
@@ -51,7 +71,7 @@ void	draw_texture(t_data *data, int x, t_line line, t_txtr *texture)
 	int		screen_y;
 	int		texture_x;
 
-	data->hit_pos = fmod(data->hit_pos / 3.79, data->map_width) / (data->map_width);
+	set_hit_position(data);
 	y = -1;
 	while (++y < line.wall_height)
 	{
