--- src/settings/presets/default.set
@@ -1,7 +1,0 @@
---default.set--
-1000
-1000
-10
-120
-0.03
-10
