+++ wow.cub
@@ -8,7 +8,7 @@ C 225,30,0
 
 
         1111111111111111111111111
-        1000100000110000000000001
+        1000100000110000000001001
         1011000001110000000000001
         1001000000000000000000001
 111111111011000001110000000000001
